test(cli): share wait_bounded between the two process tests
`init.rs` and `term_panic_order.rs` each carried a function called
`wait_bounded`, and the two bodies matched line for line: the same
`try_wait` poll, the same 20ms sleep, the same kill-then-wait-then-panic
on the deadline. Even the doc comments made the same argument, that a
named panic here beats the harness's own process timeout failing the
whole binary and naming nothing.

The only real difference was which process the two panic strings named,
so `what` is now a parameter. `term_panic_order`'s version also spawned
the `Command` itself; it spawns first and passes the `Child`, which is
what `init.rs` already did and is the narrower signature.

The `expect("poll ...")` becomes an `unwrap_or_else` that carries the I/O
error alongside the name. `expect` on a `Result<_, io::Error>` prints
both already, but the two callers' strings were fragments that only read
correctly as a prefix.

`#![cfg(unix)]` on the shared module matches both callers, which already
carry it: `commands` and `lookout` are unix only, and an integration test
file is its own compilation unit.
